fix compile error
