+++ Plugin~/WebRTCPlugin/UnityVulkanInterfaceFunctions.h
@@ -1,4 +1,7 @@
+#pragma once
+
 #include <IUnityGraphicsVulkan.h>
+#include <memory>
 
 namespace unity
 {
@@ -60,17 +63,5 @@ namespace webrtc
     private:
         T* vulkanInterface_;
     };
-
-    std::unique_ptr<UnityGraphicsVulkan> UnityGraphicsVulkan::Get(IUnityInterfaces* unityInterfaces)
-    {
-        IUnityGraphicsVulkanV2* vulkanV2 = unityInterfaces->Get<IUnityGraphicsVulkanV2>();
-        if (vulkanV2)
-            return std::make_unique<UnityGraphicsVulkanImpl<IUnityGraphicsVulkanV2>>(vulkanV2);
-        IUnityGraphicsVulkan* vulkan = unityInterfaces->Get<IUnityGraphicsVulkan>();
-        if (vulkan)
-            return std::make_unique<UnityGraphicsVulkanImpl<IUnityGraphicsVulkan>>(vulkan);
-        return nullptr;
-    }
-
 }
 }
